feat: apiのadapterクラスをMute部分とUser部分で分離
